reafac: frameのプログラムを一箇所に集めて可読性を上げる
